make ActionItems component
